Feat: 상세보기 컴포넌트 파일경로 수정
